Upgrade scripts to python 3.7.
Python 3.7 was chosen at it is the latest version of python in Debian buster

runtest, export-all were tested against nikita.oslomet.no and both
were OK. import-uttrekk was tested but failed with a problem with the
entity forfatter. This is likely due to a mismatch between the API and
the script.