Upgrade scripts to python 3.7.
Python 3.7 was chosen at it is the latest version of python in Debian buster

Note in some cases, the IDE is reporting potential issues. These have been
flaged as comments as "# Upgrade to python 3.7. Getting warning:"

runtest, export-all were tested against nikita.oslomet.no and both
were OK. import-uttrekk was tested but failed with a problem with the
entity forfatter. This is likely due to a mismatch between the API and
the script.