Remove Allows header check
It is not clear that we should be checking for the ALLOW header when testing. This was discussed in [27](https://github.com/petterreinholdtsen/noark5-tester/issues/27). Allow is part of CORS and we are likely using it wrongly. The proposed way of dealing with this is to remove it until we know what we want to achieve with it.