Remove misplaced trailing slash
During the upgrade to SB3 for nikita, it was discovered that nikita in a number of instances treated an object with a trailing slash the same as if the slash was not present. This was fixed in nikita but I see there is an instance in runtest where there is a trailing slash that I believe should not be there. This fixes the issue.