Adjusted export-xml --xml to output draft endringslogg.xml
The log is based on each entities relation to hendelseslogg, present in
Nikita.  The current Nikita implementation is not implementing the same
inheritance as is described in N5TG.

Because it is based on entity relations, entries for deleted entities
will not be included.  A different approach would be to fetch all
endringslogg/hendelseslogg entries

See also https://github.com/arkivverket/noark5-tjenestegrensesnitt-standard/pull/319 .