Adjust CORS related tests, add HEADERs with OPTION calls required by CORS.
It has been unclear if the CORS / Allows logic in runtest is correct. At some some stage
the understanding in the API spec with regards to CORS resulted in a misinterpreted
implementation. This has been partially discussed in

https://github.com/petterreinholdtsen/noark5-tester/issues/27

Other links relevantfor this change:
https://www.baeldung.com/cs/cors-preflight-requests
https://www.baeldung.com/spring-security-cors-preflight
https://fetch.spec.whatwg.org/

A CORS response should include a 'Access-Control-Allow-Methods' and a
'Access-Control-Allow-Origin'. Others may be relevant headers may be relevant
but these two are a minimum

A preflight is discussed here:

https://developer.mozilla.org/en-US/docs/Web/HTTP/Methods/OPTIONS#preflighted_requests_in_cors

In this branch we have not dealt with 'Access-Control-Request-Headers'.